core: implement `--get-dir` flag
When provided with a flag `--get-dir`, the command will instead of
writing a new log entry return the specified repository path.

This is useful for quicker access to the actual logs.

docs: updated readme with better examples on how to read and find logs.

Add an example using the `--get-dir` flag.