core: page/sub-directory support for logs
Pages are essentially sub-directories, but can be thought of as separate
pages in the journal itself.

Add Meta struct, which stores metadata related to the log, like date and
page.

Meta struct contains a function `Location()` that returns a path for the
whole log directory, including sub-directory (=page).

```
/user/someone/.caplog/capbook/page
```

Add new meta time layout, which is shown at the top of the log entry.

Add meta block for log entries, which contains information for date
with weekday and page if one was given.

Update README with new meta block example and pages.