caplog: add and use `miniflag` library instead
This minimizes the size of the source code found in caplog and also
provides a better support for help messages and flag usage in general.

This package was part of caplog previously, but was moved to its own
repository and implementation.

Removed caplog implementation of "miniflag".

Uses miniflag CommandLine Output for printing log messages.